Make all TCP communication non blocking
The receive() function parameters have been extended with boolean
`waitforanswer` which defaults to true for backward compatibility.
If set to false, the function will return instantly after each
attempt to read from the TCP socket. This allows the calling app
to determine how long to keep trying to read and the frequency of
those attemps, as well as optionally implement a break-out routine.